refactor: move app to view
